requirements.txt
flask
scikit-learn
numpy
pandas
google-cloud-vision